pyproject.toml
[tool.poetry]
name = "bedrock-agent"
version = "0.1.0"
description = ""
authors = ["Erick Okal"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
boto3 = "^1.35.87"
python-decouple = "^3.8"
python-multipart = "^0.0.20"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
